Added Generic IO
