+++ CMakeLists.txt
@@ -16,7 +16,7 @@ endif(NOT CMAKE_BUILD_TYPE)
 file(GLOB_RECURSE MY_SRC
     "${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp"
     # header don't need to be included but this might be necessary for some IDEs
-    "${CMAKE_CURRtarget_link_librariesENT_SOURCE_DIR}/src/*.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/*.h"
 )
 
 # create make target
@@ -25,7 +25,7 @@ add_executable(MolSim ${MY_SRC})
 # set cxx standard. You may raise this if you want.
 target_compile_features(MolSim
         PRIVATE
-            cxx_std_14
+            cxx_std_20
 )
 
 target_include_directories(MolSim
