+++ src/Particle.cpp
@@ -27,13 +27,13 @@ Particle::Particle(const Particle &other) {
   std::cout << "Particle generated by copy!" << std::endl;
 }
 
-// Todo: maybe use initializater list instead of copy?
+// Todo: maybe use initializer list instead of copy?
 Particle::Particle(std::array<double, 3> x_arg, std::array<double, 3> v_arg,
                    double m_arg, int type_arg) {
   x = { x_arg[0], x_arg[1], x_arg[2] };
   v = { v_arg[0], v_arg[1], v_arg[2] };
   m = m_arg;
-  type = type_arg;Eigen::Vector3d
+  type = type_arg;
   f = {0., 0., 0.};
   old_f = {0., 0., 0.};
   std::cout << "Particle generated!" << std::endl;
